+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Rename
 
 New
 * :sparkles: Displayx adds getActionBarSize() method
+* :sparkles: Bitmapx adds crop, cropTo method
 
 Behavior change:
 * :fire: Intentx createInstallAppIntent method add RequiresPermission(REQUEST_INSTALL_PACKAGES) annotated